Move printing logic back into CLIDebugHook
Since DebugPrinter previously expected a Session, anyway
